Big refactor of texture and buffer flags
Change the way we treat typed enums by adding bit test operations instead of relying on logical and.
Remove redundant enum names - BufferUsage had all its enums suffixed with Buffer, which is clear by the enum name already.

+++ code/render/coregraphics/vk/vktextureview.cc
@@ -87,7 +87,7 @@ CreateTextureView(const TextureViewCreateInfo& info)
         VK_IMAGE_USAGE_TRANSFER_DST_BIT,
         0x0 // Device exclusive?
     };
-    VkImageUsageFlags usage = Util::BitmaskConvert(info.usage, Lookup);
+    VkImageUsageFlags usage = Util::BitmaskConvert(uint(info.usage), Lookup);
 
     VkImageViewUsageCreateInfo usageInfo;
     usageInfo.pNext = nullptr;
@@ -97,7 +97,7 @@ CreateTextureView(const TextureViewCreateInfo& info)
     VkImageViewCreateInfo viewCreate =
     {
         VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO,
-        info.usage != 0x0 ? &usageInfo : nullptr,
+        uint(info.usage) != 0x0 ? &usageInfo : nullptr,
         0,
         img,
         type,
